Replace phonons tests
