--- tests/test_phonons_lammps.py
@@ -1,54 +1,0 @@
-import os
-
-from ase.build import bulk
-from phonopy.units import VaspToTHz
-import unittest
-
-from atomistics.workflows.phonons.workflow import PhonopyWorkflow
-
-try:
-    from atomistics.calculators.lammps import (
-        evaluate_with_lammps, get_potential_dataframe
-    )
-
-    skip_lammps_test = False
-except ImportError:
-    skip_lammps_test = True
-
-
-@unittest.skipIf(
-    skip_lammps_test, "LAMMPS is not installed, so the LAMMPS tests are skipped."
-)
-class TestPhonons(unittest.TestCase):
-    def test_calc_phonons(self):
-        potential = '1999--Mishin-Y--Al--LAMMPS--ipr1'
-        resource_path = os.path.join(os.path.dirname(__file__), "static", "lammps")
-        structure = bulk("Al", a=4.05, cubic=True)
-        df_pot = get_potential_dataframe(
-            structure=structure,
-            resource_path=resource_path
-        )
-        df_pot_selected = df_pot[df_pot.Name == potential].iloc[0]
-        calculator = PhonopyWorkflow(
-            structure=structure,
-            interaction_range=10,
-            factor=VaspToTHz,
-            displacement=0.01,
-            dos_mesh=20,
-            primitive_matrix=None,
-            number_of_snapshots=None,
-        )
-        structure_dict = calculator.generate_structures()
-        result_dict = evaluate_with_lammps(
-            task_dict=structure_dict,
-            potential_dataframe=df_pot_selected,
-        )
-        mesh_dict, dos_dict = calculator.analyse_structures(output_dict=result_dict)
-        self.assertEqual((324, 324), calculator.get_hesse_matrix().shape)
-        self.assertTrue('qpoints' in mesh_dict.keys())
-        self.assertTrue('weights' in mesh_dict.keys())
-        self.assertTrue('frequencies' in mesh_dict.keys())
-        self.assertTrue('eigenvectors' in mesh_dict.keys())
-        self.assertTrue('group_velocities' in mesh_dict.keys())
-        self.assertTrue('frequency_points' in dos_dict.keys())
-        self.assertTrue('total_dos' in dos_dict.keys())
